Rename the daemondiscover pkg to discover
